+++ src/shapes/collide.rs
@@ -7,7 +7,7 @@ use crate::utils::Vec3;
 
 /// Implement this trait for all objects that can be rendered in the Scene
 pub trait Collide {
-    fn get_intersection(&self, ray: &Ray, t_min: f32, t_max: f32) -> Option<HitRecord>;
+    fn get_intersection(&self, ray: &Ray, t_min: f64, t_max: f64) -> Option<HitRecord>;
     fn get_bounding_box(&self) -> Option<AABB>;
 }
 
@@ -15,7 +15,7 @@ pub trait Collide {
 pub struct HitRecord {
     pub point: Vec3,
     pub normal: Vec3,
-    pub t: f32,
+    pub t: f64,
     pub front_face: bool,
     pub material: Arc<dyn Material>,
 }
@@ -24,7 +24,7 @@ impl HitRecord {
     pub fn new(
         point: Vec3,
         outward_normal: Vec3,
-        t: f32,
+        t: f64,
         front_face: bool,
         material: Arc<dyn Material>,
     ) -> Self {
