+++ src/shapes/sphere.rs
@@ -14,13 +14,13 @@ type ArcMaterial = Arc<dyn Material + Send + Sync>;
 /// A simple Sphere to wrap in the Shape struct for rendering in a Scene.
 pub struct Sphere {
     center: Vec3,
-    radius: f32,
+    radius: f64,
     material: ArcMaterial,
 }
 
 impl Sphere {
     /// Creates a new Sphere
-    pub fn new(center: Vec3, radius: f32) -> Self {
+    pub fn new(center: Vec3, radius: f64) -> Self {
         Sphere {
             center,
             radius,
@@ -39,7 +39,7 @@ impl Sphere {
 }
 
 impl Collide for Sphere {
-    fn get_intersection(&self, ray: &Ray, t_min: f32, t_max: f32) -> Option<HitRecord> {
+    fn get_intersection(&self, ray: &Ray, t_min: f64, t_max: f64) -> Option<HitRecord> {
         let offset_center = ray.origin - self.center;
         let a = ray.direction.len_squared();
         let half_b = dot(&ray.direction, &offset_center);
